+++ pyxel/models/photon_collection/__init__.py
@@ -5,14 +5,15 @@
 #  this file, may be copied, modified, propagated, or distributed except according to
 #  the terms contained in the file ‘LICENCE.txt’.
 
-"""TBW."""
+"""Photon generation models are used to add to and manipulate data in Photon array
+inside the Detector object. """
 
 # flake8: noqa
 from .illumination import illumination
 from .load_image import load_image
 from .shot_noise import shot_noise
 from .stripe_pattern import stripe_pattern
-from .poppy import optical_psf
+from .poppy import optical_psf, optical_psf_multi_wavelength
 from .point_spread_function import load_psf, load_wavelength_psf
 from .ariel_airs import wavelength_dependence_airs
 from .aperture import simple_aperture