+++ pyxel/models/photon_collection/poppy.py
@@ -10,7 +10,7 @@
 import logging
 from collections.abc import Mapping, Sequence
 from dataclasses import dataclass
-from typing import Any, Union
+from typing import Any, Optional, Union
 
 import numpy as np
 from astropy.convolution import convolve_fft
@@ -54,6 +54,26 @@ class ThinLens:
 
     nwaves: float
     radius: float
+
+
+@dataclass
+class NewThinLens:
+    """Parameters for an idealized thin lens.
+
+    Parameters
+    ----------
+    nwaves : float
+        The number of waves of defocus, peak to valley.
+    radius : float
+        Pupil radius, in meters, over which the Zernike defocus term should be computed
+        such that rho = 1 at r = `radius`.
+    reference_wavelength : float
+        Wavelength, in meters, at which that number of waves of defocus is specified.
+    """
+
+    nwaves: float
+    radius: float
+    reference_wavelength: Optional[float] = None
 
 
 @dataclass
@@ -189,6 +209,17 @@ OpticalParameter = Union[
     ZernikeWFE,
     SineWaveWFE,
 ]
+NewOpticalParameter = Union[
+    CircularAperture,
+    NewThinLens,
+    SquareAperture,
+    RectangleAperture,
+    HexagonAperture,
+    MultiHexagonalAperture,
+    SecondaryObscuration,
+    ZernikeWFE,
+    SineWaveWFE,
+]
 
 
 def create_optical_parameter(dct: Mapping) -> OpticalParameter:
@@ -208,7 +239,73 @@ def create_optical_parameter(dct: Mapping) -> OpticalParameter:
         return CircularAperture(radius=dct["radius"])
 
     elif dct["item"] == "ThinLens":
-        return ThinLens(nwaves=dct["nwaves"], radius=dct["radius"])
+        return ThinLens(
+            nwaves=dct["nwaves"],
+            radius=dct["radius"],
+        )
+
+    elif dct["item"] == "SquareAperture":
+        return SquareAperture(size=dct["size"])
+
+    elif dct["item"] == "RectangularAperture":
+        return RectangleAperture(width=dct["width"], height=dct["height"])
+
+    elif dct["item"] == "HexagonAperture":
+        return HexagonAperture(side=dct["side"])
+
+    elif dct["item"] == "MultiHexagonalAperture":
+        return MultiHexagonalAperture(
+            side=dct["side"],
+            rings=dct["rings"],
+            gap=dct["gap"],
+        )  # cm
+
+    elif dct["item"] == "SecondaryObscuration":
+        return SecondaryObscuration(
+            secondary_radius=dct["secondary_radius"],
+            n_supports=dct["n_supports"],
+            support_width=dct["support_width"],
+        )  # cm
+
+    elif dct["item"] == "ZernikeWFE":
+        return ZernikeWFE(
+            radius=dct["radius"],
+            coefficients=dct["coefficients"],  # list of floats
+            aperture_stop=dct["aperture_stop"],
+        )  # bool
+
+    elif dct["item"] == "SineWaveWFE":
+        return SineWaveWFE(
+            spatialfreq=dct["spatialfreq"],  # 1/m
+            amplitude=dct["amplitude"],  # um
+            rotation=dct["rotation"],
+        )
+    else:
+        raise NotImplementedError
+
+
+def new_create_optical_parameter(dct: Mapping) -> NewOpticalParameter:
+    """Create a new ``OpticalParameter`` based on a dictionary.
+
+    Parameters
+    ----------
+    dct : dict
+        Dictionary to convert
+
+    Returns
+    -------
+    OpticalParameter
+        New parameters.
+    """
+    if dct["item"] == "CircularAperture":
+        return CircularAperture(radius=dct["radius"])
+
+    elif dct["item"] == "ThinLens":
+        return NewThinLens(
+            nwaves=dct["nwaves"],
+            radius=dct["radius"],
+            reference_wavelength=dct["reference_wavelength"],
+        )
 
     elif dct["item"] == "SquareAperture":
         return SquareAperture(size=dct["size"])
@@ -333,30 +430,97 @@ def create_optical_item(
         raise NotImplementedError
 
 
+def new_create_optical_item(param: NewOpticalParameter) -> "op.OpticalElement":
+    """Create a new poppy ``OpticalElement``.
+
+    Parameters
+    ----------
+    param : ``NewOpticalParameter``
+        Pyxel Optical parameters to create a poppy ``OpticalElement``.
+
+    Returns
+    -------
+    ``OpticalElement``
+        A new poppy ``OpticalElement``.
+    """
+    if isinstance(param, CircularAperture):
+        return op.CircularAperture(radius=param.radius)
+
+    elif isinstance(param, NewThinLens):
+        return op.ThinLens(
+            nwaves=param.nwaves,
+            reference_wavelength=param.reference_wavelength,
+            radius=param.radius,
+        )
+
+    elif isinstance(param, SquareAperture):
+        return op.SquareAperture(size=param.size)
+
+    elif isinstance(param, RectangleAperture):
+        return op.RectangleAperture(width=param.width, height=param.height)
+
+    elif isinstance(param, HexagonAperture):
+        return op.HexagonAperture(side=param.side)
+
+    elif isinstance(param, MultiHexagonalAperture):
+        return op.MultiHexagonAperture(
+            side=param.side, rings=param.rings, gap=param.gap
+        )
+
+    elif isinstance(param, SecondaryObscuration):
+        return op.SecondaryObscuration(
+            secondary_radius=param.secondary_radius,
+            n_supports=param.n_supports,
+            support_width=param.support_width,
+        )
+
+    elif isinstance(param, ZernikeWFE):
+        return op.ZernikeWFE(
+            radius=param.radius,
+            coefficients=param.coefficients,
+            aperture_stop=param.aperture_stop,
+        )
+
+    elif isinstance(param, SineWaveWFE):
+        return op.SineWaveWFE(
+            spatialfreq=param.spatialfreq,
+            amplitude=param.amplitude,
+            rotation=param.rotation,
+        )
+    else:
+        raise NotImplementedError(f"{param=}")
+
+
 def calc_psf(
-    wavelength: float,
+    wavelengths: Sequence[float],
     fov_arcsec: float,
     pixelscale: float,
-    optical_parameters: Sequence[OpticalParameter],
-) -> tuple[Sequence[fits.hdu.image.PrimaryHDU], Sequence["op.Wavefront"]]:
+    optical_elements: Sequence["op.OpticalElement"],
+    apply_jitter: bool = False,
+    jitter_sigma: float = 0.007,
+) -> tuple[fits.PrimaryHDU, fits.PrimaryHDU]:
     """Calculate the point spread function for the given optical system.
 
     Parameters
     ----------
-    wavelength : float
-        Wavelength of incoming light in meters.
+    wavelengths : sequence of float
+        Wavelengths of incoming light in meters.
     fov_arcsec : float, optional
         Field Of View on detector plane in arcsec.
     pixelscale : float
         Pixel scale on detector plane (arcsec/pixel).
         Defines sampling resolution of :term:`PSF`.
-    optical_parameters : list of OpticalParameter
-        List of optical parameters before detector with their specific arguments.
+    optical_elements : list of OpticalElement
+        List of optical elements to apply.
+    apply_jitter : bool
+        Defines whether jitter should be applied. Default = False.
+    jitter_sigma : float
+        Jitter sigma value in arcsec per axis, default is 0.007.
 
     Returns
     -------
-    Sequence of :term:`FITS` and sequence of Wavefront
-        Tuple of lists containing the psf and intermediate wavefronts.
+    Tuple of two :term:`FITS`
+        Tuple of psf and intermediate wavefronts.
     """
     if not WITH_POPPY:
         raise ModuleNotFoundError(
@@ -365,43 +529,102 @@ def calc_psf(
             "or 'pip install pyxel-sim[all]'"
         )
 
-    # Create the optical element(s)
-    osys = op.OpticalSystem(npix=1000)  # default: 1024
+    class PyxelInstrument(op.instrument.Instrument):
+        """Instrument class for Pyxel using poppy.instrument."""
 
-    param: OpticalParameter
-    for param in optical_parameters:
-        element: op.OpticalElement = create_optical_item(
-            param=param,
-            wavelength=wavelength,
-        )
+        def __init__(
+            self,
+            pixelscale: float,
+            optical_elements: Sequence["op.OpticalElement"],
+            fov_arcsec: float = 2,
+            name="PyxelInstrument",
+        ):
+            super().__init__(name=name)
+            self._pixelscale = pixelscale
+            self._optical_elements = optical_elements
+            self._fov_arcsec = fov_arcsec
 
-        osys.add_pupil(element)
+        def get_optical_system(
+            self,
+            fft_oversample=2,
+            detector_oversample=None,
+            fov_arcsec=None,
+            fov_pixels=None,
+            options=None,
+        ):
+            """Return an OpticalSystem instance corresponding to the instrument as currently configured.
 
-    osys.add_detector(
+            Parameters
+            ----------
+            fft_oversample : int
+                Oversampling factor for intermediate plane calculations. Default is 2
+            detector_oversample: int, optional
+                By default the detector oversampling is equal to the intermediate calculation oversampling.
+                If you wish to use a different value for the detector, set this parameter.
+                Note that if you just want images at detector pixel resolution you will achieve higher fidelity
+                by still using some oversampling (i.e. *not* setting `oversample_detector=1`) and instead rebinning
+                down the oversampled data.
+            fov_pixels : float
+                Field of view in pixels. Overrides fov_arcsec if both set.
+            fov_arcsec : float
+                Field of view, in arcseconds. Default is 2
+            options : dict
+                Other arbitrary options for optical system creation
+
+
+            Returns
+            -------
+            osys : poppy.OpticalSystem
+                an optical system instance representing the desired configuration.
+
+            """
+            osys = op.OpticalSystem(npix=1000)  # default: 1024
+
+            element: op.OpticalElement
+            for element in self._optical_elements:
+                osys.add_pupil(element)
+
+            osys.add_detector(
+                pixelscale=self._pixelscale,
+                fov_arcsec=self._fov_arcsec,
+            )
+
+            return osys
+
+    output_fits: Sequence[fits.hdu.image.PrimaryHDU]
+    wavefronts: Sequence[op.Wavefront]
+
+    # Create Instrument
+    instrument = PyxelInstrument(
+        optical_elements=optical_elements,
         pixelscale=pixelscale,
         fov_arcsec=fov_arcsec,
     )
 
-    # Calculate a monochromatic PSF
-    output_fits: Sequence[fits.hdu.image.PrimaryHDU]
-    wavefronts: Sequence[op.Wavefront]
-    output_fits, wavefronts = osys.calc_psf(
-        wavelength=wavelength,
-        return_intermediates=True,
-        normalize="last",
+    instrument.pixelscale = pixelscale
+
+    if apply_jitter:
+        instrument.options["jitter"] = "gaussian"
+        instrument.options[
+            "jitter_sigma"
+        ] = jitter_sigma  # in arcsec per axis, default 0.007
+
+    output_fits, wavefronts = instrument.calc_datacube(
+        wavelengths=wavelengths,
+        fov_arcsec=fov_arcsec,
     )
 
     return output_fits, wavefronts
 
 
-def apply_convolution(data_2d: np.ndarray, kernel_2d: np.ndarray) -> np.ndarray:
-    """Convolve an array.
+def apply_convolution(data: np.ndarray, kernel: np.ndarray) -> np.ndarray:
+    """Convolve an array in 2D or 3D.
 
     Parameters
     ----------
-    data_2d : ndarray
-        2D Array to be convolved with kernel_2d.
-    kernel_2d : ndarray
+    data : ndarray
+        2D or 3D Array to be convolved with kernel_2d.
+    kernel : ndarray
         The convolution kernel.
 
     Returns
@@ -409,16 +632,39 @@ def apply_convolution(data_2d: np.ndarray, kernel_2d: np.ndarray) -> np.ndarray:
     ndarray
         A convolved array.
     """
-    mean = np.mean(data_2d)
 
-    array_2d = convolve_fft(
-        data_2d,
-        kernel=kernel_2d,
+    if kernel.ndim == 2:
+        mean = np.mean(data)
+    elif kernel.ndim == 3:
+        integrated = kernel.sum(axis=0)
+        mean = integrated.mean()
+    else:
+        raise ValueError
+
+    *_, num_rows, num_cols = kernel.shape
+
+    assert num_rows == num_cols
+    # resize kernel, if kernel size too big.
+    if num_rows > 10:
+        import skimage.transform as sk
+
+        if kernel.ndim == 2:
+            new_shape: tuple[int, ...] = (10, 10)
+        elif kernel.ndim == 3:
+            num_wavelengths, _, _ = kernel.shape
+            new_shape = num_wavelengths, 10, 10
+
+        resized_kernel = sk.resize(kernel, output_shape=new_shape, anti_aliasing=False)
+        kernel = resized_kernel / resized_kernel.sum()
+
+    array = convolve_fft(
+        data,
+        kernel=kernel,
         boundary="fill",
         fill_value=mean,
     )
 
-    return array_2d
+    return array
 
 
 def optical_psf(
@@ -426,6 +672,8 @@ def optical_psf(
     wavelength: float,
     fov_arcsec: float,
     optical_system: Sequence[Mapping[str, Any]],
+    apply_jitter: bool = False,
+    jitter_sigma: float = 0.007,
 ) -> None:
     """Model function for poppy optics model: convolve photon array with psf.
 
@@ -439,6 +687,10 @@ def optical_psf(
         Field Of View on detector plane in arcsec.
     optical_system : list of dict
         List of optical elements before detector with their specific arguments.
+    apply_jitter : bool
+        Defines whether jitter should be applied. Default = False.
+    jitter_sigma : float
+        Jitter sigma value in arcsec per axis, default is 0.007.
     """
     logging.getLogger("poppy").setLevel(
         logging.WARNING
@@ -457,22 +709,135 @@ def optical_psf(
         create_optical_parameter(dct) for dct in optical_system
     ]
 
+    optical_elements: Sequence["op.OpticalElement"] = [
+        create_optical_item(param=param, wavelength=wavelength)
+        for param in optical_parameters
+    ]
+
     # Processing
     # Get a Point Spread Function
-    images, wavefronts = calc_psf(
-        wavelength=wavelength,
+    image_hdu_3d: fits.PrimaryHDU
+    # wavefront_hdu_3d: fits.PrimaryHDU
+    image_hdu_3d, wavefront_3d = calc_psf(
+        wavelengths=[wavelength],
         fov_arcsec=fov_arcsec,
         pixelscale=detector.geometry.pixel_scale,
-        optical_parameters=optical_parameters,
+        optical_elements=optical_elements,
+        apply_jitter=apply_jitter,
+        jitter_sigma=jitter_sigma,
     )
 
-    # Extract 'first_image'
-    first_image, *other_images = images
+    data_3d: np.ndarray = image_hdu_3d.data
+    data_2d: np.ndarray = data_3d[0, :, :]
 
     # Convolution
     new_array_2d: np.ndarray = apply_convolution(
-        data_2d=detector.photon.array,
-        kernel_2d=first_image.data,
+        data=detector.photon.array,
+        kernel=data_2d,
     )
 
     detector.photon.array = new_array_2d
+
+
+def optical_psf_multi_wavelength(
+    detector: Detector,
+    wavelengths: tuple[float, float],
+    fov_arcsec: float,
+    optical_system: Sequence[Mapping[str, Any]],
+    apply_jitter: bool = False,
+    jitter_sigma: float = 0.007,
+) -> None:
+    """Model function for poppy optics model: convolve photon array with psf.
+
+    Parameters
+    ----------
+    detector : Detector
+        Pyxel Detector object.
+    wavelengths : tuple of floats
+        Wavelengths of incoming light in meters.
+    fov_arcsec : float
+        Field Of View on detector plane in arcsec.
+    optical_system : list of dict
+        List of optical elements before detector with their specific arguments.
+    apply_jitter : bool
+        Defines whether jitter should be applied. Default = False.
+    jitter_sigma : float
+        Jitter sigma value in arcsec per axis, default is 0.007.
+    """
+    import xarray as xr
+    from astropy.units import Quantity
+
+    logging.getLogger("poppy").setLevel(
+        logging.WARNING
+    )  # TODO: Fix this. See issue #81
+
+    # Validation and Conversion stage
+    # These steps will be probably moved into the YAML engine
+    if len(wavelengths) != 2:
+        raise ValueError("Expecting two wavelengths in parameter 'wavelengths'.")
+
+    min_wavelength, max_wavelength = wavelengths
+    if min_wavelength <= 0:
+        raise ValueError("Expecting strictly positive value for the 'wavelengths'")
+
+    if min_wavelength > max_wavelength:
+        raise ValueError(
+            f"Min wavelength must be smaller that max wavelength. Got: {wavelengths!r}"
+        )
+
+    if fov_arcsec < 0.0 or detector.geometry.pixel_scale < 0.0:
+        raise ValueError(
+            "Expecting strictly positive value for " "'fov_arcsec' and 'pixel_scale'."
+        )
+
+    # Convert 'optical_system' to 'optical_parameters'
+    optical_parameters: Sequence[NewOpticalParameter] = [
+        new_create_optical_parameter(dct) for dct in optical_system
+    ]
+
+    optical_elements: Sequence["op.OpticalElement"] = [
+        new_create_optical_item(param=param) for param in optical_parameters
+    ]
+
+    # Get current wavelengths (in nm)
+    start_wavelength = Quantity(min_wavelength, unit="m")
+    end_wavelength = Quantity(max_wavelength, unit="m")
+    wavelengths_nm: Quantity = Quantity(
+        detector.photon.array_3d["wavelength"], unit="nm"
+    )
+
+    tolerance = Quantity(1e-7, unit="m")
+    selected_wavelengths_nm: Quantity = wavelengths_nm[
+        np.logical_and(
+            wavelengths_nm >= (start_wavelength - tolerance),
+            wavelengths_nm <= (end_wavelength + tolerance),
+        )
+    ]
+    if selected_wavelengths_nm.size == 0:
+        raise ValueError
+
+    # Processing
+    # Get a Point Spread Function
+    image_3d: fits.PrimaryHDU
+    wavefront_3d: fits.PrimaryHDU
+    image_3d, wavefront_3d = calc_psf(
+        wavelengths=selected_wavelengths_nm.to("m").value,
+        fov_arcsec=fov_arcsec,
+        pixelscale=detector.geometry.pixel_scale,
+        optical_elements=optical_elements,
+        apply_jitter=apply_jitter,
+        jitter_sigma=jitter_sigma,
+    )
+
+    # Convolution
+    new_array_3d: np.ndarray = apply_convolution(
+        data=detector.photon.array_3d.to_numpy(),
+        kernel=image_3d.data,
+    )
+
+    array_3d = xr.DataArray(
+        new_array_3d,
+        dims=["wavelength", "y", "x"],
+        coords={"wavelength": selected_wavelengths_nm.value},
+    )
+    detector.photon.array_3d = array_3d