+++ pyxel/notebook/__init__.py
@@ -15,6 +15,7 @@ from .jupyxel import (
     display_array,
     display_detector,
     display_persist,
+    display_scene,
 )
 from .html_representation import display_html